+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 - Bump minimum Python version from 3.9 to 3.9.2.
-- `nitrokey.nk3.updates`: Remove `UpdatePath`, `get_extra_information` and `get_finalization_wait_retries` from public API.
+- `nitrokey.nk3.updates`:
+  - Remove `UpdatePath`, `get_extra_information` and `get_finalization_wait_retries` from public API.
+  - Show warning when updating from bootloader mode and add `confirm_update_from_bootloader` method to `UpdateUi`.
 - Add support for updates to Nitrokey 3 firmware v1.8.2.
 
 [All Changes](https://github.com/Nitrokey/nitrokey-sdk-py/compare/v0.2.4...HEAD)